+++ src/REvoDesign/shortcuts/wrappers/ligand_converters.py
@@ -22,7 +22,6 @@ wrapped_smiles_conformer_single = registry.register(
     "smiles_conformer_single",
     shortcut_smiles_conformer_single,
     use_thread=True,
-    use_progressbar=True,
 )
 
 # Register the SMILES Conformer (Batch)
@@ -30,7 +29,6 @@ wrapped_smiles_conformer_batch = registry.register(
     "smiles_conformer_batch",
     shortcut_smiles_conformer_batch,
     use_thread=True,
-    use_progressbar=True,
 )
 
 # Register the SDF to Rosetta Parameter Conversion
@@ -38,5 +36,4 @@ wrapper_sdf2rosetta_params = registry.register(
     "sdf2rosetta_params",
     shortcut_sdf2rosetta_params,
     use_thread=True,
-    use_progressbar=True,
 )