+++ tests/tools/customized_widgets/test_dialog_wrapper_registry.py
@@ -46,7 +46,7 @@ def test_dialog_wrapper_registry_register_with_thread(mock_partial, mock_load_ya
     registry.funcs = {}  # reset
 
     kwargs = {"key": "value"}
-    registry.register("test_func", dummy_function, use_thread=True, use_progressbar=False, kwargs=kwargs)
+    registry.register("test_func", dummy_function, use_thread=True, kwargs=kwargs)
 
     assert "test_func" in registry.funcs
     mock_partial.assert_called_once_with(run_wrapped_func_in_thread, dummy_function, **kwargs)
@@ -63,7 +63,7 @@ def test_dialog_wrapper_registry_register_without_thread(mock_load_yaml, mock_ex
     registry = DialogWrapperRegistry("test_category")
     registry.funcs = {}  # reset
 
-    registry.register("test_func", dummy_function, use_progressbar=False, use_thread=False)
+    registry.register("test_func", dummy_function, use_thread=False)
 
     assert registry.funcs["test_func"] == dummy_function
 
@@ -91,7 +91,7 @@ def test_dialog_wrapper_registry_window_wrapper_dynamic_values(mock_load_yaml, m
     registry = DialogWrapperRegistry("test_category")
     assert registry.funcs == {}
 
-    wrapper = registry.register("test_func", dummy_function, use_progressbar=False, has_dynamic_values=True)
+    wrapper = registry.register("test_func", dummy_function, has_dynamic_values=True)
     assert registry.funcs == {"test_func": dummy_function}
 
     assert wrapper.__doc__ is not None
@@ -108,7 +108,7 @@ def test_dialog_wrapper_registry_window_wrapper_no_dynamic_values(mock_load_yaml
 
     registry = DialogWrapperRegistry("test_category")
 
-    wrapper = registry.register("test_func", dummy_function, use_progressbar=False, has_dynamic_values=False)
+    wrapper = registry.register("test_func", dummy_function, has_dynamic_values=False)
 
     assert wrapper.__doc__ is not None
     assert "no dynamic values" in wrapper.__doc__
@@ -123,7 +123,7 @@ def test_dialog_wrapper_registry_unregister(mock_load_yaml, mock_exists, mock_ya
         return
 
     registry = DialogWrapperRegistry("test_category")
-    registry.register("test_func", dummy_function, use_progressbar=False)
+    registry.register("test_func", dummy_function)
 
     assert registry.funcs == {"test_func": dummy_function}